Adding a bunch of performance improvements
This commit adds a series of improvements, either by making the
code (slightly) smaller, or by storing function calls and lookups
as termporary variables.

I also removed the transition on the `.spot` element, since that
was making the overall look-and-feel of the shadow movement slower.

I made `showIndicators` default to falsey rather than truthy, since
it seemed more like a development-debug tool and not really something
desired with a finished product.

I changed several of the events that were `onmousemove` to
`addEventListener`, matching closer to the listeners `load`, `resize`,
and `scroll`.

I also changed the mousemove events, removing the anonymous function
wrapper (which is more costly) and replacing with the single `initiateSpot`
event function.

I also removed a bunch of unused variables.